requirements.txt
numpy
pandas
missingno
matplotlib.pyplot
catboost
morfeusz2